requirements.txt
Django
djangorestframework
psycopg2
PyYAML
djangorestframework-simplejwt
drf-yasg
whitenoise